Migrate timer to a web worker.
