+++ src/App.vue
@@ -1,11 +1,21 @@
 <script setup>
+import { onMounted, ref } from 'vue';
 import Stopwatch from './components/Stopwatch.vue';
+
+const isSupported = ref(false);
+
+onMounted(() => {
+  isSupported.value = typeof(Worker) !== "undefined";
+});
 
 </script>
 
 <template>
   <div>
-    <stopwatch></stopwatch>
+    <stopwatch v-if="isSupported"></stopwatch>
+    <div v-else>
+      <p>Sorry, your browser does not support this application</p>
+    </div>
   </div>
 </template>
